+++ src/db/mongoose.js
@@ -57,18 +57,18 @@ mongoose.connect('mongodb://127.0.0.1:27017/task-manager-api', {
 //     console.log(error)
 // })
 
-const Task = mongoose.model('tasks', {
-    description: {
-        type: String,  // Booleans, dates ,array, binary data, object IDs and more
-        required: true,
-        trim:true
+// const Task = mongoose.model('tasks', {
+//     description: {
+//         type: String,  // Booleans, dates ,array, binary data, object IDs and more
+//         required: true,
+//         trim:true
         
-    },
-    completed: {
-        type: Boolean,
-        default:false
-    }
-})
+//     },
+//     completed: {
+//         type: Boolean,
+//         default:false
+//     }
+// })
 
 // const newTask = new Task({
 //     description: "Clean the car",
